requirements.txt
requests>=2.0.0
markdownify>=0.11.0
beautifulsoup4>=4.0.0
trafilatura>=1.6.0
tqdm>=4.0.0
google-genai>=0.1.0
python-dotenv>=1.0.0
pymupdf4llm>=0.0.17
pymupdf>=1.23.0
